+++ CHANGELOG.md
@@ -3,13 +3,19 @@
 ## Unreleased
 
 ### Added
+-Added new `enchanted-fab` component with support for badges and AI icons, accompanied by comprehensive unit tests, snapshots, and Storybook documentation.
+- Added story for `enchanted-badge`
 
 ### Fixed
 - Fixed `enchanted-circular-progress` indeterminant variant style and story.
 
 ### Changed
+- Enhanced `enchanted-badge` to improve accessibility and styling consistency.
+- Added new property `color, border, badge, text` for `enchanted-badge`
 
 ### Breaking changes
+- Enable badge="dot" in opensearch for filter
+- Enable badge="text" text=${this.count} for advance search
 
 ## 3.0.0